extraction des requetes au serveur dans une classe spécifique
